# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DiffEntry": {
        "properties": {
          "new": {
            "title": "New"
          },
          "old": {
            "title": "Old"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "title": "DiffEntry",
        "type": "object"
      },
...
      "DriverRequest": {
        "properties": {
          "base_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Revision"
          }
        },
        "title": "DriverRequest",
        "type": "object"
      },
      "DriverResult": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "legacy_switch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legacy Switch"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "ok"
        ],
        "title": "DriverResult",
        "type": "object"
      },
...
      "EntityOption": {
        "properties": {
          "device_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Class"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "entity_id",
          "name",
          "state",
          "unit",
          "device_class"
        ],
        "title": "EntityOption",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobInfo": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "grace_s": {
            "title": "Grace S",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_finished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Finished"
          },
          "last_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Outcome"
          },
          "last_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run Id"
          },
          "last_started": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Started"
          },
          "next_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run"
          },
          "paused": {
            "title": "Paused",
            "type": "boolean"
          },
          "running": {
            "title": "Running",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "trigger",
          "paused",
          "running",
          "next_run",
          "last_started",
          "last_finished",
          "last_outcome",
          "last_run_id",
          "grace_s"
        ],
        "title": "JobInfo",
        "type": "object"
      },
...
      "LegacyApplyRequest": {
        "properties": {
          "base_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Revision"
          }
        },
        "title": "LegacyApplyRequest",
        "type": "object"
      },
...
      "MarketReconcileRequest": {
        "properties": {
          "force_end": {
            "default": false,
            "title": "Force End",
            "type": "boolean"
          },
          "trigger_entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Entity"
          }
        },
        "title": "MarketReconcileRequest",
        "type": "object"
      },
      "MarketSession": {
        "properties": {
          "direction": {
            "title": "Direction",
            "type": "string"
          },
          "end_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Reason"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "power_w": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power W"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "direction",
          "source",
          "mode",
          "power_w",
          "started_at",
          "updated_at"
        ],
        "title": "MarketSession",
        "type": "object"
      },
...
      "MlRequest": {
        "properties": {
          "historic_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Historic Days"
          },
          "n_trials": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Trials"
          },
          "sklearn_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sklearn Model"
          }
        },
        "title": "MlRequest",
        "type": "object"
      },
...
      "RunStarted": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobInfo"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "job",
          "run_id"
        ],
        "title": "RunStarted",
        "type": "object"
      },
...
      "SaveResponse": {
        "properties": {
          "diff": {
            "items": {
              "$ref": "#/components/schemas/DiffEntry"
            },
            "title": "Diff",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "diff"
        ],
        "title": "SaveResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EMHASS Lens",
    "version": "schema"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/driver/hand-back": {
      "post": {
        "description": "Put EMHASS Lens in dry run and turn the HACS integration's Auto MPC back on.",
        "operationId": "hand_back_api_driver_hand_back_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hand Back",
        "tags": [
          "system"
        ]
      }
    },
    "/api/driver/take-over": {
      "post": {
        "description": "Turn the HACS integration's Auto MPC off and make EMHASS Lens drive EMHASS (live, Auto MPC on).",
        "operationId": "take_over_api_driver_take_over_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Take Over",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/ha/entities": {
      "get": {
        "description": "Entities for the settings pickers (Home Assistant's states are re-read at most every 30 s). Only for\nrequests that may change settings: the optional direct port has no login and mustn't expose every state.",
        "operationId": "ha_entities_api_ha_entities_get",
        "parameters": [
          {
            "description": "Comma-separated domains, e.g. sensor,input_number",
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated domains, e.g. sensor,input_number",
              "title": "Domain"
            }
          },
          {
            "description": "Words to find in the entity id or name; best matches first",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Words to find in the entity id or name; best matches first",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntityOption"
                  },
                  "title": "Response Ha Entities Api Ha Entities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ha Entities",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/legacy/import": {
      "post": {
        "operationId": "legacy_apply_api_legacy_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Content"
          }
        },
        "summary": "Legacy Apply",
        "tags": [
          "system"
        ]
      }
    },
...
    "/api/market/reconcile": {
      "post": {
        "description": "Reconcile now (live: acts; force_end ends the open session whatever the command says).",
        "operationId": "market_reconcile_api_market_reconcile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/MarketReconcileRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Reconcile",
        "tags": [
          "system"
        ]
      }
    },
    "/api/market/sessions": {
      "get": {
        "operationId": "market_sessions_api_market_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MarketSession"
                  },
                  "title": "Response Market Sessions Api Market Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market Sessions",
        "tags": [
          "system"
        ]
      }
    },
    "/api/ml/{action}": {
      "post": {
        "operationId": "ml_action_api_ml__action__post",
        "parameters": [
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ml Action",
        "tags": [
          "system"
        ]
      }
    },
...
  }
}